hotfix

+++ system.cpp
@@ -206,7 +206,7 @@ void System::tick(int clk) {
         if (full_system && w_addr >= UART_LITE_BASE && w_addr < UART_LITE_BASE+0x1000) { /* UART Lite */
             if (top->m_axi_wstrb == 0xF0) w_addr += 4;
             else if (top->m_axi_wstrb != 0) {
-                cerr << "Write request with unsupported strobe value (" << std::dec << top->m_axi_wstrb << ")" << endl;
+                cerr << "Write request with unsupported strobe value (" << std::dec << (int)(top->m_axi_wstrb) << ")" << endl;
                 Verilated::gotFinish(true);
             }
             if (w_addr == UART_LITE_BASE + 4*UART_LITE_REG_TXFIFO) cout << (char)(top->m_axi_wdata >> 32) << std::flush;
